requirements.txt
ariadne>=0.12.0
pydantic>=1.7.3
pyjwt>=1.7.1
jinjasql>=0.1.8
httpx>=0.16.1
sentry_sdk>=0.19.5
tortoise-orm>=0.16.19,<0.18.0
asyncpg>=0.21.0
ariadne>=0.12.0
async_lru>=1.0.2
aioredis>=1.3.1,<2.0.0
newrelic>=5.22.1.152
pytest>=6.2.1
python-dotenv>=0.15.0
asynctest>=0.13.0
arq>=0.19.1
Jinja2>=2.5,<3.0
markupsafe==2.0.1
